+++ SourceCode/TopologicalWatershed/TopologicalWatershed.cpp
@@ -7,7 +7,7 @@
 int main (void)
 {
     // Building Component Tree for dummy array
-    vector<unsigned int> dummyPixelArray = { 4,5,6,8,3,4,9,1,7 };
+    vector<unsigned short> dummyPixelArray = { 4,5,6,8,3,4,9,1,7 };
     ImageFourNeighborType image(dummyPixelArray,3,3);
     ComponentTree tree(image);
